move ProductKey and getProductConfiguration to LabKeySiteWrapper
